// pop-web/src/app/shared/openapi/model/vitals-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Measure } from './measure';


export interface VitalsCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who\'s vitals are assesed
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the vitals were recorded.
     */
    date: string;
    height?: Measure | null;
    weight?: Measure | null;
    bloodPressureSystolic?: Measure | null;
    bloodPressureDiastolic?: Measure | null;
    temperature?: Measure | null;
}